Create main.cpp
Task 1 – Number Guessing Game (C++)
A fun and interactive game where the player tries to guess a randomly generated number between 1 and 100. The program gives hints if the guess is too high or too low, helping the player improve their logic and problem-solving skills. This project was completed as part of the CodeAlpha Internship to practice C++ fundamentals and logical thinking.